backend/requirements.txt
fastapi
uvicorn
pydantic
openai
google-generativeai
python-dotenv
tenacity
